service classes made

+++ src/main/java/repositories/AlbumRepository.java
@@ -3,5 +3,5 @@ package repositories;
 import entities.Albums;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AlbumRepository extends CrudRepository<Albums, Integer> {
+public interface AlbumRepository extends CrudRepository<Albums, Long> {
 }
