+++ src/main/java/repositories/ArtistsRepository.java
@@ -3,5 +3,5 @@ package repositories;
 import entities.Artists;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ArtistsRepository extends CrudRepository<Artists, Integer> {
+public interface ArtistsRepository extends CrudRepository<Artists, Long> {
 }
